First commit of the new codebase
The system works in play&editor mode of Unity. But there are some
problem with the native-managed communication. If you start/stop
multiple senders at the same time Unity Receivers can receive
nullpointer as resourceview and loose the ability to display the
texture. If you start a new instance from the same receiver app this new
instance works .So two apps who listen to the same shareName are getting
different resourceView pointers.