Unity5 update: scene cleanup
Added DXGI_FORMAT_B8G8R8A8_UNORM as texture format option to Unity
SpoutSender and in native plugin.
SpoutSender now checks if the camera has a targetTexture and auto adds
the current texture if needed.
Added a ReadMe.txt for a little introduction.
Cleanup of the demo scenes.
Changed the include path for c++ files to the default Spout2
installation path